feat: move router level to ts
